+++ plugin.video.orftvthek/addon.xml
@@ -1,10 +1,11 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<addon id="plugin.video.orftvthek" name="ORF TVthek" version="0.10.3" provider-name="sofaking">
+<addon id="plugin.video.orftvthek" name="ORF TVthek" version="0.12.0" provider-name="sofaking">
     <requires>
         <import addon="xbmc.python" version="2.14.0"/>
-        <import addon="script.module.future" version="0.16.0"/>
-        <import addon="script.module.kodi-six" version="0.0.5"/>
-        <import addon="script.module.simplejson" version="2.0.10" optional="false"/>
+        <import addon="script.module.future" version="0.17.1"/>
+        <import addon="script.module.kodi-six" version="0.0.7"/>
+        <import addon="script.module.simplejson" version="3.16.1"/>
+        <import addon="script.module.inputstreamhelper" optional="true"/>
     </requires>
     <extension point="xbmc.python.pluginsource" library="default.py">
         <provides>video</provides>
@@ -12,14 +13,21 @@
     <extension point="xbmc.addon.metadata">
         <platform>all</platform>
         <language>de en</language>
-        <summary lang="de">ORF TVthek</summary>
-        <summary lang="en">ORF TVthek</summary>
-        <description lang="de">ORF TVthek - Ermöglicht Ihnen den Zugriff auf die ORF TVthek Video Platform</description>
-        <description lang="en">ORF TVthek - This plugin provides access to the Austrian "ORF TVthek"</description>
+        <summary lang="de_DE">ORF TVthek</summary>
+        <summary lang="en_GB">ORF TVthek</summary>
+        <description lang="de_DE">ORF TVthek - Ermöglicht Ihnen den Zugriff auf die ORF TVthek Video Platform</description>
+        <description lang="en_GB">ORF TVthek - This plugin provides access to the Austrian "ORF TVthek"</description>
         <license>GPL-2.0-only</license>
         <forum>https://forum.kodi.tv/showthread.php?tid=159835</forum>
         <email>sofaking@gettingmoney.at</email>
         <website>https://tvthek.orf.at</website>
         <source>https://github.com/s0faking/plugin.video.orftvthek</source>
+        <assets>
+            <icon>resources/icon.png</icon>
+            <fanart>resources/fanart.jpg</fanart>
+        </assets>
+        <news>v0.12.0 (01/05/2021)
+            [feature] drm livestreams
+        </news>
     </extension>
 </addon>
