+++ plugin.video.orftvthek/resources/lib/Helpers.py
@@ -61,5 +61,5 @@ def parameters_string_to_dict(parameters):
 
 
 def debugLog(message, loglevel=xbmc.LOGDEBUG):
-    output = "[ORF TVTHEK] " + py2_encode(message)
-    #xbmc.log(msg=output, level=loglevel)
+    output = py2_encode(message)
+    xbmc.log(msg=output, level=loglevel)
